# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "REVENUE HUB",
        "description": "Below is the API endpoint for the Revenue Hub",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/property": {
            "get": {
                "tags": [
                    "Property"
                ],
                "summary": "Get all properties",
                "operationId": "673e4f505d2feda35deb1c326e67f146",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PropertyResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No properties found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status",
                                            "type": "string"
                                        },
                                        "message": {
                                            "description": "Message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Property"
                ],
                "summary": "Add New Property Details",
                "description": "This allow staff admin to add new property details",
                "operationId": "addProperty",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "pid": {
                                                "type": "integer"
                                            },
                                            "occupant": {
                                                "type": "string"
                                            },
                                            "prop_addr": {
                                                "type": "string"
                                            },
                                            "street_name": {
                                                "type": "string"
                                            },
                                            "asset_no": {
                                                "type": "string"
                                            },
                                            "cadastral_zone": {
                                                "type": "string"
                                            },
                                            "prop_type": {
                                                "type": "string"
                                            },
                                            "prop_use": {
                                                "type": "string"
                                            },
                                            "rating_dist": {
                                                "type": "string"
                                            },
                                            "annual_value": {
                                                "type": "string"
                                            },
                                            "rate_payable": {
                                                "type": "string"
                                            },
                                            "grand_total": {
                                                "type": "string"
                                            },
                                            "category": {
                                                "type": "string"
                                            },
                                            "group": {
                                                "type": "string"
                                            },
                                            "active": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "pid": "integer",
                                    "occupant": "string",
                                    "prop_addr": "string",
                                    "street_name": "string",
                                    "asset_no": "string",
                                    "cadastral_zone": "string",
                                    "prop_type": "string",
                                    "prop_use": "string",
                                    "rating_dist": "string",
                                    "annual_value": "string",
                                    "rate_payable": "string",
                                    "grand_total": "string",
                                    "category": "string",
                                    "group": "string",
                                    "active": "string"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Property added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PropertyResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "All Fields are Required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "All Fields are required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "An error occured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An error occured"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PropertyResource": {
                "title": "Store Property Resource",
                "description": "Store Property Resource",
                "type": "object",
                "example": {
                    "pid": 3176683,
                    "occupant": "THE OCCUPIER",
                    "prop_addr": "540 Kaci Extensions Suite 724 Francoborough, MS 22188",
                    "street_name": "46329 McDermott Courts Apt. 058",
                    "asset_no": "AMC/B14/TR/9",
                    "cadastral_zone": "30769-6233",
                    "prop_type": "SPLITTED",
                    "prop_use": "COMMERCIAL",
                    "rating_dist": "APO",
                    "annual_value": 200000,
                    "rate_payable": 45000,
                    "grand_total": 4500000,
                    "category": "RESIDENTIAL",
                    "group": "AMAC2",
                    "active": "ACTIVE",
                    "created_at": "2024-05-06T11:54:58.000000Z",
                    "updated_at": "2024-05-06T11:54:58.000000Z"
                }
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer token)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}